Correct Setting of LibreSN:
decodeSerialNumber is moved to LibreUtils.
It is used also for manual scanning.
We set LibreSN for tomato (a bug fix) and also for manual scanning.
