+++ app/src/main/java/com/eveningoutpost/dexdrip/UtilityModels/Blukon.java
@@ -358,7 +358,7 @@ private static final int POSITION_OF_SENSOR_STATUS_BYTE = 17;
             */
 
             if (JoH.pratelimit(BLUKON_DECODE_SERIAL_TIMER, GET_DECODE_SERIAL_DELAY)) {
-                String SensorSn = decodeSerialNumber(buffer);
+                String SensorSn = LibreUtils.decodeSerialNumber(buffer);
                 // TODO: Only write this after checksum was verified
                 PersistentStore.setString("LibreSN", SensorSn);
             }
@@ -622,43 +622,6 @@ private static final int POSITION_OF_SENSOR_STATUS_BYTE = 17;
         } else {
             currentCommand = "";
         }
-    }
-
-    // This function assumes that the UID is starting at place 3, and is 8 bytes long
-    public static String decodeSerialNumber(byte[] input) {
-        
-        byte[] uuid = new byte[]{0, 0, 0, 0, 0, 0, 0, 0};
-        String lookupTable[] =
-                {
-                        "0", "1", "2", "3", "4", "5", "6", "7", "8", "9",
-                        "A", "C", "D", "E", "F", "G", "H", "J", "K", "L",
-                        "M", "N", "P", "Q", "R", "T", "U", "V", "W", "X",
-                        "Y", "Z"
-                };
-        byte[] uuidShort = new byte[]{0, 0, 0, 0, 0, 0, 0, 0};
-        int i;
-
-        for (i = 2; i < 8; i++) uuidShort[i - 2] = input[(2 + 8) - i];
-        uuidShort[6] = 0x00;
-        uuidShort[7] = 0x00;
-
-        String binary = "";
-        String binS = "";
-        for (i = 0; i < 8; i++) {
-            binS = String.format("%8s", Integer.toBinaryString(uuidShort[i] & 0xFF)).replace(' ', '0');
-            binary += binS;
-        }
-        
-        String v = "0";
-        char[] pozS = {0, 0, 0, 0, 0};
-        for (i = 0; i < 10; i++) {
-            for (int k = 0; k < 5; k++) pozS[k] = binary.charAt((5 * i) + k);
-            int value = (pozS[0] - '0') * 16 + (pozS[1] - '0') * 8 + (pozS[2] - '0') * 4 + (pozS[3] - '0') * 2 + (pozS[4] - '0') * 1;
-            v += lookupTable[value];
-        }
-        Log.e(TAG, "decodeSerialNumber=" + v);
-
-        return v;
     }
